hrtimer: Ignore hrtimer requeue when not in CB.
This patch removes a BUG_ON in the hrtimer root
callback. Allows the hook to be called when
not actually in the callback state. A BUG_ON
forbid this state when a timer needed to be
rearmed. Instead, we just ignore the rearm
gracefully.

This feature is needed for budget enforcement
under bandwidth-inheritance.