Add new TASK_DAEMON release policy.
Special policy for daemon tasks. We want daemon tasks
to run in the background automatically (hence, early
releasing behavior), but daemon tasks may not be able
to signal when their jobs have completed (because they're
closed source). Thus, we need a hybrid of the TASK_SPORADIC
(which automatically handles new job releases) and TASK_EARLY.