litmus_lock(), etc.: Fail early if callee not RT
This patch updates the litmus_lock() and similar
system calls to fail early if the callees are not
real-time tasks. This may have to change in the future,
if there were a locking protocol that supported
both general purpose and litmus tasks. However,
no such protocol exists at the moment, so let's
go with what makes sense for now.