R2DGLP: Promote PQ over another's donor to FQ.
The R2DGLP has the optimization where, when a resource
holder WITHOUT a donor releases a replica, the donor
of ANOTHER task is enqueued on the the replica's fifo
queue. However, this should only happen if the highest-
priority request in the PQ is not among the top-m-highest
priority blocked requests. This patch enforces this
condition.

Note, however, that in 2012/2013, we thought this case
was impossible to arrise. That is, the highest-priority
request in PQ would never be among the top-m-highest
priority blocked requests. However, we can no longer
recall our reasoning behind this assertion. Emperical
testing suggests that our 2012/2013 thinking was
correct---the case where a request in PQ is promoted
to an FQ over a donor never seems to fire.

Thus, if we were correct in 2012/2013, this patch
adds just a tiny bit of extra overhead. If we were
incorrect, then this patch fixes that bug.